requirements.txt
fastapi
uvicorn
mysql-connector-python
pandas
numpy
scikit-learn
tensorflow
transformers
torch
newsapi-python
python-dotenv
